[CHUNGCHUN-111] 링크 수정, 삭제 API (#32)
* 링크 수정 API 추가

* 링크 수정 api 추가 및 수정, 삭제 인가 작업

* 주석 제거

* 공통된 문자열 템플릿 함수화

+++ src/main/java/org/example/youth_be/user/service/UserService.java
@@ -5,6 +5,7 @@ import org.example.youth_be.artwork.repository.ArtworkRepository;
 import org.example.youth_be.artwork.service.request.ArtworkPaginationRequest;
 import org.example.youth_be.common.CursorPagingCommon;
 import org.example.youth_be.common.PageResponse;
+import org.example.youth_be.common.exceptions.YouthBadRequestException;
 import org.example.youth_be.common.exceptions.YouthDuplicateException;
 import org.example.youth_be.common.exceptions.YouthNotFoundException;
 import org.example.youth_be.common.jwt.MainAccessTokenProvider;
@@ -15,10 +16,7 @@ import org.example.youth_be.user.domain.UserLinkEntity;
 import org.example.youth_be.user.enums.UserRoleEnum;
 import org.example.youth_be.user.repository.UserLinkRepository;
 import org.example.youth_be.user.repository.UserRepository;
-import org.example.youth_be.user.service.request.UserAdditionSignupRequest;
-import org.example.youth_be.user.service.request.UserSignupRequest;
-import org.example.youth_be.user.service.request.LinkRequest;
-import org.example.youth_be.user.service.request.UserProfileUpdateRequest;
+import org.example.youth_be.user.service.request.*;
 import org.example.youth_be.artwork.service.response.ArtworkResponse;
 import org.example.youth_be.user.service.response.UserMyInformation;
 import org.example.youth_be.user.service.response.UserMyPage;
@@ -30,6 +28,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
+import static org.example.youth_be.common.util.DebuggingTemplate.NotAuthorized;
 
 @Service
 public class UserService {
@@ -77,7 +77,10 @@ public class UserService {
     }
 
     @Transactional
-    public Long createUserLink(Long userId, LinkRequest request) {
+    public Long createUserLink(Long userId, LinkRequest request, TokenClaim claim) {
+        if (claim.isNotAuthorized(userId)) {
+            throw new YouthBadRequestException("권한이 없는 사용자입니다.", NotAuthorized(userId, claim));
+        }
         userRepository.findById(userId).orElseThrow(() -> new YouthNotFoundException("해당 ID의 유저를 찾을 수 없습니다.", null));
         UserLinkEntity newUserLinkEntity = UserLinkEntity.builder()
                 .userId(userId)
@@ -89,11 +92,22 @@ public class UserService {
     }
 
     // 링크 수정
-    // 이미지 삭제
+    @Transactional
+    public Long updateUserLink(Long userId, Long linkId, UserLinkUpdateRequest request, TokenClaim claim) {
+        if (claim.isNotAuthorized(userId)) {
+            throw new YouthBadRequestException("권한이 없는 사용자입니다.", NotAuthorized(userId, claim));
+        }
+        UserLinkEntity userLinkEntity = userLinkRepository.findByIdAndUserId(linkId, userId).orElseThrow(() -> new YouthNotFoundException("링크를 찾을 수 없습니다", null));
+        userLinkEntity.updateLink(request.getTitle(), request.getUrl());
+        return userLinkEntity.getId();
+    }
 
     @Transactional
-    public void deleteUserLink(Long userId, Long linkId) {
-        userLinkRepository.findByIdAndUserId(linkId, userId).orElseThrow(() -> new YouthNotFoundException("링크를 찾을 수 없습니다", null));
+    public void deleteUserLink(Long userId, Long linkId, TokenClaim claim) {
+        if (claim.isNotAuthorized(userId)) {
+            throw new YouthBadRequestException("권한이 없는 사용자입니다.", NotAuthorized(userId, claim));
+        }
+        userLinkRepository.findByIdAndUserId(linkId, claim.getUserId()).orElseThrow(() -> new YouthNotFoundException("링크를 찾을 수 없습니다", null));
         userLinkRepository.deleteById(linkId);
     }
